refactor: update command structure and clean up unused files
- Removed the deprecated test script for task management.
- Refactored command handling in OpenDevCoach.CLI.Commands to use a new TaskCommands module.
- Updated command syntax to use slashes (e.g., /help) for better clarity.
- Added temporary library path to .gitignore for context management.
